# api/openapi.yaml
openapi: 3.0.3
info:
  title: Lux Trading API
  description: |
    One surface for every market a Lux interface shows.

    A market on a Lux chain is quoted from our own contracts. A market on any
    other chain is quoted from the pools Uniswap deployed there, read directly
    over that chain's JSON-RPC — no vendor API, no upstream account, nothing
    that stops working when somebody says no. The response is the same shape
    either way, because a client should not have to know which arm answered it.

    Every path lives under /v1/trade. api.lux.cloud carries the chain at
    /v1/chain, the C-chain at the root and other services besides, so this is
    one namespace with one owner rather than a handful of generic names on a
    shared host.

    Reads are public and unauthenticated. Nothing here signs or sends a
    transaction: the swap, approval and position paths return unsigned calldata
    for a wallet to sign.
  version: 3.0.0
  contact:
    name: Lux Industries
    url: https://lux.exchange
servers:
  - url: https://api.lux.cloud
    description: Production
  - url: http://localhost:8080
    description: A local gateway
...
paths:
  /v1/trade/venues:
    get:
      operationId: venues
      tags: [reference]
      summary: The chains this deployment can price, and what it reads them from
      description: |
        Ask this first. A venue is a contract on a chain, so the answer is
        grouped by chain: `native` marks a chain we settle ourselves through the
        Lux precompiles, and the rest are read from their own pools.

        Without this a client discovers a chain is unquotable by quoting on it
        and reading an error.
      parameters:
        - name: chainId
          in: query
          required: false
          schema: { type: integer, format: int64 }
          description: Restrict the answer to one chain.
      responses:
        '200':
          description: One entry per chain, in ascending chain id.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ChainVenues' }
        '404':
          description: No venue here reads that chain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /v1/trade/quote:
    post:
      operationId: quote
      tags: [quote]
      summary: The best price across every venue that holds the pair
      description: |
        Our own chains are asked first, because they are the venue — an upstream
        opinion about a market we settle ourselves is a second opinion. A venue
        holding no liquidity returns nothing rather than an error, so the next
        source is asked in turn.

        `providerName` and `route[].poolType` name the venue that answered.
        A quote is a reading of a pool at a moment and is cached for seconds.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuoteRequest' }
      responses:
        '200':
          description: The best price, and where it came from.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Quote' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: |
            No venue here holds this pair. On a chain this deployment reads,
            that is a fact about the pools and not about a missing upstream
            account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /v1/trade/quotes:
    post:
      operationId: quotes
      tags: [quote]
      summary: Every venue's price for the pair, best first
      description: |
        The same question as /v1/trade/quote without the last step, from the
        same source, so the head of this list is what that path returns.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuoteRequest' }
      responses:
        '200':
          description: Every venue that holds the pair, in descending amountOut.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Quote' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: No venue here holds this pair.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /v1/trade/route:
    post:
      operationId: route
      tags: [quote]
      summary: Multi-hop paths through the pools, ranked
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteRequest' }
      responses:
        '200':
          description: Candidate routes, best output first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/trade/swap:
    post:
      operationId: swap
      tags: [swap]
      summary: An unsigned transaction for the best route
      description: |
        Quotes the pair, then returns calldata for a wallet to sign. Nothing
        here holds a key or sends anything.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SwapRequest' }
      responses:
        '200':
          description: An unsigned transaction.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/UnsignedTx' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/trade/approval/check:
    post:
      operationId: checkApproval
      tags: [allowance]
      summary: What the spender may already move
      description: |
        Reads `allowance(owner, spender)` on the token, over the same endpoint
        the pool is quoted from. `approveTx` is returned only when an approval
        is actually needed — handed back beside "you already approved" it is a
        transaction somebody sends, and the user pays gas for a state change
        that changes nothing.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllowanceRequest' }
      responses:
        '200':
          description: The allowance the chain reports, and whether it covers the amount.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Approval' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502':
          description: |
            The chain could not be read, or this deployment does not read it.
            Never a guessed number.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /v1/trade/approval/build:
    post:
      operationId: buildApproval
      tags: [allowance]
      summary: An unsigned ERC-20 approve
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllowanceRequest' }
      responses:
        '200':
          description: An unsigned transaction.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/UnsignedTx' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /v1/trade/permit2/check:
    post:
      operationId: checkPermit2
      tags: [allowance]
      summary: Whether a live Permit2 grant covers the amount
      description: |
        Permit2 answers `(amount, expiration, nonce)`, and the expiration is
        half the answer: a grant that has run out is not a grant. A
        `signatureRequest` comes back only when one is needed.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Permit2Request' }
      responses:
        '200':
          description: The Permit2 allowance, and whether it is live and sufficient.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Permit2' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '502':
          description: The chain could not be read, or this deployment does not read it.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /v1/trade/permit2/build:
    post:
      operationId: buildPermit2
      tags: [allowance]
      summary: A Permit2 grant, as a transaction or as typed data to sign
      parameters:
        - name: mode
          in: query
          schema: { type: string, enum: [tx, sig], default: tx }
          description: |
            `tx` returns an unsigned Permit2 approve; `sig` returns the EIP-712
            PermitSingle for the wallet to sign instead.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Permit2Request' }
      responses:
        '200':
          description: An unsigned transaction, or an EIP-712 signature request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /v1/trade/pools:
    get:
      operationId: pools
      tags: [liquidity]
      summary: Pools on a chain
      parameters:
        - name: chainId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: token0
          in: query
          schema: { type: string }
        - name: token1
          in: query
          schema: { type: string }
        - name: protocol
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: The pools that answered.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
...
components:
...
  schemas:
    Envelope:
      type: object
      description: Every answer carries the same envelope.
      required: [success]
      properties:
        success: { type: boolean }
        data: {}
        error:
          type: string
          description: Present only when success is false.
        requestId: { type: string }
...
    ChainVenues:
      type: object
      required: [chainId, native, venues]
      properties:
        chainId: { type: integer, format: int64 }
        native:
          type: boolean
          description: True where the chain is settled here through the Lux precompiles.
        venues:
          type: array
          items: { $ref: '#/components/schemas/Venue' }
...